🚚 Storybook preview 파일 ts -> tsx 로 수정
